+++ .github/logs/CHANGELOG.md
@@ -22,3 +22,4 @@ and animated the arrow in the Thingamajig JEI screen
 updated it in the gradle.properties file +
 Reminder to check the changelog for more information
 - Oop, sorry, forgot to update the changelog for the commit
+- Added maxTransfer (filling with shift) support to the bowyery JEI support